Restored header files for  individual algorithms
Realised that the header files for  individual algorithms are needed. Putting them back. If required we can move them to interface(.h) and src(.cc) areas.
