+++ SimTracker/SiPhase2Digitizer/plugins/PixelDigitizerAlgorithm.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <cmath>
 
-#include "SimTracker/SiPhase2Digitizer/plugins/Phase2TrackerDigitizerAlgorithm.h"
+#include "SimTracker/SiPhase2Digitizer/plugins/PixelDigitizerAlgorithm.h"
 #include "SimDataFormats/TrackingHit/interface/PSimHitContainer.h"
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
@@ -18,31 +18,6 @@
 #include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
 #include "Geometry/CommonTopologies/interface/PixelTopology.h"
-
-class PixelDigitizerAlgorithm : public Phase2TrackerDigitizerAlgorithm {
-public:
-  PixelDigitizerAlgorithm(const edm::ParameterSet& conf);
-  ~PixelDigitizerAlgorithm() override;
-
-  // initialization that cannot be done in the constructor
-  void init(const edm::EventSetup& es) override;
-
-  // void initializeEvent();
-  // run the algorithm to digitize a single det
-  void accumulateSimHits(const std::vector<PSimHit>::const_iterator inputBegin,
-                         const std::vector<PSimHit>::const_iterator inputEnd,
-                         const size_t inputBeginGlobalIndex,
-                         const uint32_t tofBin,
-                         const Phase2TrackerGeomDetUnit* pixdet,
-                         const GlobalVector& bfield) override;
-  void add_cross_talk(const Phase2TrackerGeomDetUnit* pixdet) override;
-
-  // Addition four xtalk-related parameters to PixelDigitizerAlgorithm specific parameters initialized in Phase2TrackerDigitizerAlgorithm
-  const double odd_row_interchannelCoupling_next_row_;
-  const double even_row_interchannelCoupling_next_row_;
-  const double odd_column_interchannelCoupling_next_column_;
-  const double even_column_interchannelCoupling_next_column_;
-};
 
 using namespace edm;
 using namespace sipixelobjects;
