+++ SimTracker/SiPhase2Digitizer/plugins/SSDigitizerAlgorithm.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <cmath>
 
-#include "SimTracker/SiPhase2Digitizer/plugins/Phase2TrackerDigitizerAlgorithm.h"
+#include "SimTracker/SiPhase2Digitizer/plugins/SSDigitizerAlgorithm.h"
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
@@ -10,23 +10,6 @@
 // Geometry
 #include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
-
-class SSDigitizerAlgorithm : public Phase2TrackerDigitizerAlgorithm {
-public:
-  SSDigitizerAlgorithm(const edm::ParameterSet& conf);
-  ~SSDigitizerAlgorithm() override;
-
-  // initialization that cannot be done in the constructor
-  void init(const edm::EventSetup& es) override;
-
-  // run the algorithm to digitize a single det
-  void accumulateSimHits(const std::vector<PSimHit>::const_iterator inputBegin,
-                         const std::vector<PSimHit>::const_iterator inputEnd,
-                         const size_t inputBeginGlobalIndex,
-                         const uint32_t tofBin,
-                         const Phase2TrackerGeomDetUnit* pixdet,
-                         const GlobalVector& bfield) override;
-};
 
 using namespace edm;
